chore: Rename address-bar
